backend/requirements-dev.txt
# EBR Platform - Development Dependencies
# Install with: pip install -r requirements.txt -r requirements-dev.txt

-r requirements.txt

# =============================================================================
# Django Development Tools
# =============================================================================
django-extensions==3.2.3
Werkzeug==3.1.3

# =============================================================================
# Testing
# =============================================================================
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
Faker==33.1.0

# =============================================================================
# Code Quality
# =============================================================================
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0
django-stubs==5.1.1
djangorestframework-stubs==3.15.1
